Extract, visualize, and save DEG lists

ProcessDEGs wraps several plotting functions to generate figures specifically for the differentially expressed genes between each possible comparison. It is called by RunDESeq2 but can also be re-run with the RunDESeq2 output or its own output if you want to save time and don't need to generate all of the exploratory data analysis figures again.

ProcessDEGs(dds, rld, vsd, outpath, level, plot.annos,
  padj.thresh = 0.05, fc.thresh = 2, plot.box = TRUE, top.n = 100)

Arguments

dds
A DESeqDataSet object as returned by DESeq.

rld
A RangedSummarizedExperiment object of rlog transformed counts as returned by RunDESeq2.

vsd
A RangedSummarizedExperiment object of vst transformed counts as returned by RunDESeq2.

outpath
Path to directory to be used for output.

level
String defining variable of interest.

plot.annos
String or character vector defining the column(s) in samplesheet to use to annotate figures.

padj.thresh
Number or numeric scalar indicating the adjusted p-value cutoff(s) to be used for determining "significant" differential expression. If multiple are given, multiple tables/plots will be generated using all combinations of padj.thresh and fc.thresh.

fc.thresh
Number or numeric scalar indicating the log2 fold-change cutoff(s) to be used for determining "significant" differential expression. If multiple are given, multiple tables/plots will be generated using all combinations of padj.thresh and fc.thresh.

plot.box
Boolean indicating whether box plots for DEGs should be created for each comparison. If so, the top.n genes will be plotted.

top.n
Number of differentially expressed genes to create boxplots for, ranked by adj. p-value after applying padj.thresh and fc.thresh thresholds. If multiple thresholds are provided, the lowest fold-change and highest adj. p-value thresholds will be used.

Value

Named List containing a list named 'res.list' containing DESeqResults objects for all comparisons generated by ProcessDEGs, a DESeqDataSet object named 'dds' from running DESeq, a RangedSummarizedExperiment object named 'rld' from running rlog, and a RangedSummarizedExperiment object from running vst named 'vsd'.

Details

This function will generate many figures in addition to saving the counts and results.