Run DESeq2 analyses

This function performs a high-level differential gene expression analysis with DESeq2.

RunDESeq2(outpath, quants.path, samplesheet, tx2gene, level,
  padj.thresh = 0.05, fc.thresh = 2, plot.annos = NULL,
  plot.box = TRUE, top.n = 100, block = NULL, count.filt = 10)

Arguments

outpath
Path to directory to be used for output. Additional directories will be generated within this folder.

quants.path
Path to directory containing a directory for each sample with a salmon-generated quant.sf file inside.

samplesheet
Path to samplesheet containing sample metadata.

tx2gene
Path to file with transcript IDs in first column and gene identifiers in second. Used for gene-level summarization.

level
String defining variable of interest.

padj.thresh
Number or numeric scalar indicating the adjusted p-value cutoff(s) to be used for determining "significant" differential expression. If multiple are given, multiple tables/plots will be generated using all combinations of padj.thresh and fc.thresh.

fc.thresh
Number or numeric scalar indicating the log2 fold-change cutoff(s) to be used for determining "significant" differential expression. If multiple are given, multiple tables/plots will be generated using all combinations of padj.thresh and fc.thresh.

plot.annos
String or character vector defining the column(s) in samplesheet to use to annotate figures.

plot.box
Boolean indicating whether box plots for DEGs should be created for each comparison. If so, the top.n genes will be plotted.

top.n
Number of differentially expressed genes to create boxplots for, ranked by adj. p-value after applying padj.thresh and fc.thresh thresholds. If multiple thresholds are provided, the lowest fold-change and highest adj. p-value thresholds will be used.

block
String or character vector defining the column(s) in samplesheet to use to block for unwanted variance, e.g. batch or technical effects.

count.filt
Number indicating read threshold. Genes with fewer counts than this number summed across all samples will be removed from the analysis.

Value

Named List containing a DESeqDataSet object from running DESeq, a RangedSummarizedExperiment object from running rlog, and a RangedSummarizedExperiment object from running vst.

See also